+++ src/applications/hca/tests/config/medicarePartAEffectiveDate.unit.spec.js
@@ -8,7 +8,7 @@ import {
   DefinitionTester,
   submitForm,
   getFormDOM,
-} from 'platform/testing/unit/schemaform-utils.jsx';
+} from 'platform/testing/unit/schemaform-utils';
 import formConfig from '../../config/form';
 
 describe('Hca medicare', () => {
