+++ src/applications/hca/tests/config/serviceInformation.unit.spec.js
@@ -8,7 +8,7 @@ import {
   DefinitionTester,
   fillData,
   fillDate,
-} from 'platform/testing/unit/schemaform-utils.jsx';
+} from 'platform/testing/unit/schemaform-utils';
 import formConfig from '../../config/form';
 
 describe('Hca serviceInformation', () => {
